# perf: updated fetchPriority and sizes for avatars