--- a/packages/shared/src/components/ProfilePicture.tsx
+++ b/packages/shared/src/components/ProfilePicture.tsx
@@ -82,6 +82,7 @@ export function setOnError(
 const resizeSrcReplaceRule: Record<string, (src: string) => string> = {
   [SocialProvider.Google]: (src) => src.replace(/s96-c$/, 's64-c'),
   [SocialProvider.GitHub]: (src) => {
+    return src;
     const search = new URLSearchParams(src);
     search.set('s', '64');
     return src.replace(/\?.+$/, `?${search.toString()}`);
@@ -89,12 +90,11 @@ const resizeSrcReplaceRule: Record<string, (src: string) => string> = {
 };
 
 const getSocialProviderFromSrc = (src: string): SocialProvider | null => {
-  const { host } = new URL(src);
-  if (host.includes('googleusercontent.')) {
+  if (src.includes('googleusercontent.')) {
     return SocialProvider.Google;
   }
 
-  if (host.includes('githubusercontent.')) {
+  if (src.includes('githubusercontent.')) {
     return SocialProvider.GitHub;
   }
 
@@ -107,6 +107,7 @@ const getResizedSrc = (src: string) => {
     return src;
   }
 
+  console.log({ src, provider });
   const getReplacedSrc = resizeSrcReplaceRule[provider];
   return getReplacedSrc?.(src) ?? src;
 };
@@ -143,6 +144,8 @@ function ProfilePictureComponent(
       ProfileImageSize.Small,
       ProfileImageSize.XSmall,
     ].includes(size);
+
+  console.log({ user });
 
   const imageAlt = `${user.username || user.name || user.id}'s profile`;
   const imageSrc = isImageResizable ? getResizedSrc(user.image) : user.image;