--- a/packages/shared/src/components/post/NewComment.tsx
+++ b/packages/shared/src/components/post/NewComment.tsx
@@ -139,6 +139,12 @@ function NewCommentComponent(
             pictureClasses,
             getProfilePictureClasses(ProfileImageSize.Large),
           )}
+          aria-hidden
+          fetchPriority="low"
+          height={40}
+          loading="lazy"
+          role="presentation"
+          width={40}
         />
       )}
       <span className="text-text-tertiary">Share your thoughts</span>