## Changes
- Lazy load for avatar in comment box;
- Reduced Google/Github avatar sizes in comments;
- Used _t_logo_ tranformation for guest users (from 512px to 48px size);
- Lazy loaded and low fetch priority for comment avatars

### Before - LCP 3.2s / Blocking time 5.690s
<img width="966" alt="image" src="https://github.com/user-attachments/assets/2c1cee45-c9a2-4fa8-a6e7-d8c7509a05c1">

### After  - LCP 2.2s / Blocking time 5.420s
<img width="966" alt="image" src="https://github.com/user-attachments/assets/1c5d1be9-c83f-433e-a9df-14152f729940">


## Events

Did you introduce any new tracking events?

<!--
If yes please remove the comment HTML comment tags and fill the table below

Don't forget to update the [Analytics Taxonomy sheet](https://docs.google.com/spreadsheets/d/18Lv7zXges9QfVX5VYL1a-Hyl0e1sQ3sLr0OK8YZWKXI/edit#gid=0)

Log the new/changed events below:

| Type   | event_name  | value |
|--------|-------------|-------|
| Change/New | event name  | extra: { ... } |

-->

## Experiment

Did you introduce any new experiments?

<!--
If yes please remove the comment HTML comment tags and follow the instructions below

Don't forget to send a message to the [#experiments](https://dailydotdev.slack.com/archives/C02JAUF8HJL/p1715175315620999) channel, following the template in slack, and adding a link to the message here.

> [!IMPORTANT]
> Please do not merge the PR until the experiment enrolment is approved.

-->

## Manual Testing

> [!CAUTION]
> Please make sure existing components are not breaking/affected by this PR

<!--
If relevant, please remove the comment HTML comment tags and fill the checkboxes below

### On those affected packages:
- [ ] Have you done sanity checks in the webapp?
- [ ] Have you done sanity checks in the extension?
- [ ] Does this not break anything in companion?

### Did you test the modified components media queries?
- [ ] MobileL (420px)
- [ ] Tablet (656px)
- [ ] Laptop (1020px)

#### Did you test on actual mobile devices?
- [ ] iOS (Chrome and Safari)
- [ ] Android

-->

<!--
Copy and paste the below line outside the HTML comment tags to link this PR to the ticket in Jira

AS-{number} #done
or
MI-{number} #done
-->


### Preview domain
https://perf-comment-avatar.preview.app.daily.dev/posts/the-wordpress-ecosystem-has-lost-its-mind--j992wvxiw