@ilasw 2024-10-25
Also I've noticed a collateral performance increase in feed, now even there we have resized GitHub/Google avatars:

### Before - 377KB
<img width="1241" alt="image" src="https://github.com/user-attachments/assets/ca92899d-5e3b-4590-b100-b163ea9b356f">


### After - 12KB
<img width="1241" alt="image" src="https://github.com/user-attachments/assets/e2d711f9-fc84-4606-8730-586684b3004b">